Conflicts:
	bin/FileHandler.class

+++ src/Peach.java
@@ -1,7 +1,7 @@
 import java.awt.Color;
 
 public class Peach extends GameObject {
-	public Peach(int x, int y, int h, int w) {
+	public Peach(float x, float y, int h, int w) {
 		super(x, y, h, w);
 		isSolid = false;
 		name = "peach";
