+++ src/Platform.java
@@ -3,7 +3,7 @@ import java.awt.Color;
 public class Platform extends GameObject{
 	private boolean hasLadder = false;
 	
-	public Platform(int x, int y, int h, int w) {
+	public Platform(float x, float y, int h, int w) {
 		super(x, y, h, w);
 		color = Color.red;
 		isSolid = true;
