+++ src/Powerup.java
@@ -1,6 +1,6 @@
 
 public class Powerup extends GameObject {
-	public Powerup(int x, int y, int h, int w) {
+	public Powerup(float x, float y, int h, int w) {
 		super(x, y, h, w);
 		isSolid = false;
 		name = "powerup";
